auditor/requirements.txt
requests==2.32.*
psycopg2-binary==2.9.*
